+++ src/index.ts
@@ -1,4 +1,8 @@
+// project info
 import { name, version } from "../package.json";
+
+// abstracted server RCON methods
+import * as server from "./server";
 
 // constants
 const emojiNumbers = [
@@ -14,163 +18,120 @@ const emojiNumbers = [
 	"9️⃣",
 ];
 
-// internal
-let connectionAttempts: number = 0;
-let serverData = {
-	onlinePlayerNames: [] as string[],
-	onlinePlayerCount: 0,
-	maxPlayerCount: 0,
-};
-
 // message
 console.info(`Starting (${name} - v${version})... `);
 
 /**
- * RCON
+ * RCON Methods
  */
-
 import Rcon from "rcon-srcds";
 
-// init RCON client
-const rconClient = new Rcon({
-	host: process.env.RCON_HOST as string,
-	port: Number(process.env.RCON_PORT),
-});
-
 // run `list` command, and format response
-const getPlayers = async () => {
+const getPlayers = async (serverInfo: IServerInfo) => {
 	// get player list data
-	let playerListResponse = await rconClient
-		.execute("list")
-		.then((response) => {
-			return response as string;
-		})
-		.catch((error) => {
-			console.error(error);
-			return;
-		});
+	let playerListResponse = await server[serverInfo.TYPE].getList(
+		serverInfo.rconClient
+	);
 
 	// no response
 	if (!playerListResponse) return;
 
 	// init new server data
-	let newServerData = {
-		onlinePlayerNames: [] as string[],
-		onlinePlayerCount: 0,
-		maxPlayerCount: 0,
-	};
-
-	// format response and separate player numbers/names
-	let numberDataUnparsed, playerNamesUnparsed;
-	[numberDataUnparsed, playerNamesUnparsed] = playerListResponse
-		.replace("There are ", "")
-		.replace(" of a max of", "")
-		.replace(" players online", "")
-		.split(":");
-	newServerData.onlinePlayerNames = playerNamesUnparsed
-		.trim()
-		.replaceAll(",", "")
-		.split(" ");
-
-	// get player counts
-	[newServerData.onlinePlayerCount, newServerData.maxPlayerCount] =
-		numberDataUnparsed.split(" ").map((string: string) => {
-			return parseInt(string);
-		});
+	let newServerData =
+		server[serverInfo.TYPE].parseDataFromList(playerListResponse);
 
 	// update server data and discord presence
-	if (newServerData !== serverData) {
-		serverData = newServerData;
-		updateDiscordPresence();
+	if (newServerData !== serverInfo.data) {
+		serverInfo.data = newServerData;
+		updateDiscordPresence(serverInfo);
 	}
 };
 
 // authenticate
-const authRCONClient = async () => {
+const authRCONClient = async (serverInfo: IServerInfo) => {
 	// broadcast
-	if (connectionAttempts <= 3)
+	if (serverInfo.connectionAttempts <= 3)
 		console.info(
 			"Attempting to establish RCON connection..." +
-				(connectionAttempts > 1
-					? ` (Attempt ${connectionAttempts})`
+				(serverInfo.connectionAttempts > 1
+					? ` (Attempt ${serverInfo.connectionAttempts})`
 					: "")
 		);
-	else if (connectionAttempts == 4)
+	else if (serverInfo.connectionAttempts == 4)
 		console.warn(
 			"Failed to connect after 3 attempts! Will continue to attempt reconnecting in 5 minute intervals from now on..."
 		);
 
 	// log attempt
-	connectionAttempts += 1;
+	serverInfo.connectionAttempts += 1;
 
 	// attempt auth
-	await rconClient
-		.authenticate(process.env.RCON_PASSWORD as string)
+	await serverInfo.rconClient
+		.authenticate(serverInfo.RCON_PASSWORD as string)
 		.then(() => {
 			// broadcast
 			console.info("RCON connection established!");
 
 			// clear attempts
-			connectionAttempts = 0;
+			serverInfo.connectionAttempts = 0;
 		});
 
 	// fetch
-	fetchData();
+	fetchData(serverInfo);
 };
 
 // establish RCON connection and begin data fetching
-const fetchData = async () => {
+const fetchData = async (serverInfo: IServerInfo) => {
 	// check if authenticated
-	if (!rconClient.connected) {
+	if (!serverInfo.rconClient.connected) {
 		// already tried to connect 3 or more times- try again in 5 minutes
-		if (connectionAttempts > 3) {
+		if (serverInfo.connectionAttempts > 3) {
 			setTimeout(authRCONClient, 300000);
 		}
 
 		// retry RCON connection immediately
 		else {
-			authRCONClient();
+			authRCONClient(serverInfo);
 		}
 		return;
 	}
 
 	// get players
-	await getPlayers();
+	await getPlayers(serverInfo);
 
 	// queue next check (3 seconds)
-	setTimeout(fetchData, 3000);
+	setTimeout(() => {
+		fetchData(serverInfo);
+	}, 3000);
 };
 
-// start RCON connection and data fetching
-fetchData();
-
 /**
- * Discord Bot
+ * Discord Methods
  */
-
 import { ActivityType, Client, Events, GatewayIntentBits } from "discord.js";
 
-// create discord client
-const discordClient = new Client({ intents: [GatewayIntentBits.Guilds] });
-
 // update presence
-const updateDiscordPresence = () => {
+const updateDiscordPresence = (serverInfo: IServerInfo) => {
 	// init state
-	let state = rconClient.connected ? `${emojiNumbers[0]} Online` : "OFFLINE";
+	let state = serverInfo.rconClient.connected
+		? `${emojiNumbers[0]} Online`
+		: "OFFLINE";
 
 	// update state
-	if (rconClient.connected) {
-		if (serverData.onlinePlayerCount > 0)
-			state = `${[...serverData.onlinePlayerCount.toString()]
+	if (serverInfo.rconClient.connected) {
+		if (serverInfo.data.onlinePlayerCount > 0)
+			state = `${[...serverInfo.data.onlinePlayerCount.toString()]
 				.map((number) => {
 					return emojiNumbers[parseInt(number)];
 				})
-				.join("")} Online (${serverData.onlinePlayerNames.join(", ")})`;
+				.join("")} Online (${serverInfo.data.onlinePlayerNames.join(
+				", "
+			)})`;
 	}
 
 	// set presence
-	discordClient.user?.setPresence({
-		status: rconClient.connected ? "online" : "dnd",
+	serverInfo.discordClient.user?.setPresence({
+		status: serverInfo.rconClient.connected ? "online" : "dnd",
 		activities: [
 			{
 				type: ActivityType.Custom,
@@ -181,14 +142,58 @@ const updateDiscordPresence = () => {
 	});
 };
 
-// events
-discordClient.once(Events.ClientReady, (client) => {
-	// broadcast
-	console.info(`Discord Bot (${client.user.username}) Connected!`);
+// get config
+import config from "../config.json";
 
-	// update presence
-	updateDiscordPresence();
-});
+// get servers
+const servers = config.servers as IServerConfig[];
 
-// login client
-discordClient.login(process.env.DISCORD_BOT_TOKEN);
+// create discord and RCON client for each server
+for (const server of servers) {
+	/**
+	 * RCON
+	 */
+
+	// init RCON client
+	const rconClient = new Rcon({
+		host: server.RCON_HOST,
+		port: server.RCON_PORT,
+	});
+
+	/**
+	 * Discord
+	 *  */
+
+	// create discord client
+	const discordClient = new Client({ intents: [GatewayIntentBits.Guilds] });
+
+	// create new server info var
+	let serverInfo = {
+		...server,
+		...{
+			rconClient,
+			discordClient,
+			connectionAttempts: 0,
+			data: {
+				onlinePlayerNames: [] as string[],
+				onlinePlayerCount: 0,
+				maxPlayerCount: 0,
+			},
+		},
+	} as IServerInfo;
+
+	// set up discord bot events
+	discordClient.once(Events.ClientReady, (client) => {
+		// broadcast
+		console.info(`Discord Bot (${client.user.username}) Connected!`);
+
+		// update presence
+		updateDiscordPresence(serverInfo);
+	});
+
+	// login client
+	await discordClient.login(serverInfo.DISCORD_BOT_TOKEN);
+
+	// start RCON connection and data fetching
+	fetchData(serverInfo);
+}