+++ src/server/index.ts
@@ -1,0 +1,2 @@
+export { default as minecraft } from "./minecraft";
+export { default as palworld } from "./palworld";