+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   simultaneous bot and server instances support.
+-   support for minecraft and palworld game servers.
+
 ## [0.1.0] - 2024-01-19
 
 ### Added